# backend/boxen/assets/openapi.yaml
openapi: 3.1.2
...
info:
  title: Boxen Local API
  version: 1.0.0
  summary: Same-origin API for the locally hosted Boxen inventory system.
  description: >-
    This contract is normative for Boxen v1 HTTP field names and wire formats.
    The API is served only from the configured local Boxen origin: trusted-LAN
    HTTP is supported in development; production requires HTTPS.
    Anonymous access is configured server-side as off, viewer, or editor (default).
    When enabled, GET /session issues a per-browser session cookie and CSRF token
    without login. Other protected operations still require that session cookie.
    Anonymous access never grants user management, system administration, backup,
    or permanent purge permissions.
servers:
  - url: /api/v1
    description: Same-origin local Boxen server
...
security:
  - cookieSession: []
paths:
...
  /session:
    get:
      tags: [Authentication]
      operationId: getSession
      summary: Read the current user, CSRF token, and capabilities.
      description: >-
        Preserves a valid local or anonymous session. When anonymous access is enabled,
        a missing or expired session is replaced with a new per-browser anonymous
        session at the configured viewer or editor role. When disabled, a valid
        local user session is required. An explicit Origin must match the configured
        origin when anonymous access is enabled.
      security:
        - cookieSession: []
        - {}
      responses:
        '200':
          description: Current session.
          headers:
            Set-Cookie:
              description: Per-browser session cookie when an anonymous session is created.
              schema: { type: string }
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionView'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: List local users. Owner only.
      responses:
        '200':
          description: Local users.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/UserView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Users]
      operationId: createUser
      summary: Create a local user. Owner only.
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserCreate' }
      responses:
        '201':
          description: User created.
          headers:
            Location: { schema: { type: string } }
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
...
  /boxes:
    get:
      tags: [Boxes]
      operationId: listBoxes
      summary: List boxes visible to the current user.
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/TagFilter'
        - $ref: '#/components/parameters/CollectionFilter'
        - name: lifecycle
          in: query
          schema:
            type: string
            enum: [active, archived, all]
            default: active
        - name: sort
          in: query
          schema:
            type: string
            enum: [updated_desc, name_asc, created_desc]
            default: updated_desc
      responses:
        '200':
          description: Page of boxes.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BoxPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
    post:
      tags: [Boxes]
      operationId: createBox
      summary: Create a box. Editor or owner.
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BoxCreate' }
      responses:
        '201':
          description: Created box.
          headers:
            Location: { schema: { type: string } }
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BoxDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
...
  /boxes/{box_code}/images:
    parameters:
      - $ref: '#/components/parameters/BoxCode'
    get:
      tags: [Images]
      operationId: listBoxImages
      summary: List non-deleted images for a box.
      responses:
        '200':
          description: Ordered image summaries.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/ImageView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Images]
      operationId: uploadBoxImage
      summary: Upload and finalize one image. Editor or owner.
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                caption:
                  type: string
                  maxLength: 500
      responses:
        '201':
          description: Ready image.
          headers:
            Location: { schema: { type: string } }
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImageView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '507': { $ref: '#/components/responses/InsufficientStorage' }
...
  /boxes/{box_code}/items:
    parameters:
      - $ref: '#/components/parameters/BoxCode'
    get:
      tags: [Inventory]
      operationId: listInventoryItems
      summary: List confirmed inventory for a box.
      parameters:
        - name: include_removed
          in: query
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Inventory items.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/InventoryItemView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Inventory]
      operationId: createInventoryItem
      summary: Create a confirmed manual inventory item. Editor or owner.
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InventoryItemCreate' }
      responses:
        '201':
          description: Created item.
          headers:
            Location: { schema: { type: string } }
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InventoryItemView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
...
  /boxes/{box_code}/observations:
    get:
      tags: [Analysis]
      operationId: listBoxObservations
      summary: List AI observations for a box.
      parameters:
        - $ref: '#/components/parameters/BoxCode'
        - name: decision
          in: query
          schema:
            type: string
            enum: [pending, accepted, rejected, superseded, all]
            default: pending
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Observation page.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ObservationPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /search:
    get:
      tags: [Discovery]
      operationId: searchBoxes
      summary: Search box code, name, description, confirmed contents, tags, and collection names.
      parameters:
        - $ref: '#/components/parameters/TagFilter'
        - $ref: '#/components/parameters/CollectionFilter'
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 256 }
        - name: include_archived
          in: query
          schema: { type: boolean, default: false }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Ranked search page.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SearchPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
  /labels.pdf:
    post:
      tags: [Labels]
      operationId: renderLabelSheet
      summary: Render one label per selected box on US Letter sheets. Editor or owner.
      description: Read-only PDF generation. Pin order is preserved; collections use name/code order and include archived members. No partial output for missing boxes or empty collections. Maximum 500 labels.
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LabelSheetRequest' }
      responses:
        '200':
          description: Deterministic Letter PDF, ten 4 x 2-inch labels per page.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Content-Disposition:
              schema: { type: string }
          content:
            application/pdf:
              schema: { type: string, format: binary }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /system:
    get:
      tags: [Operations]
      operationId: getSystemStatus
      summary: Read authorized component, capacity, and model status.
      responses:
        '200':
          description: System status.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SystemStatusView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: boxen_session
      description: >-
        HttpOnly, SameSite=Strict session cookie, Secure over HTTPS. Issued by
        local login, first-owner setup, or GET /session when anonymous access is
        enabled. Anonymous mutations retain the same Origin, CSRF, role, and
        resource-version checks as local user sessions.
  headers:
    RequestId:
      description: Correlation identifier for this request.
      schema: { type: string, minLength: 8, maxLength: 128 }
    ETag:
      description: Strong resource version identifier.
      schema: { type: string }
  parameters:
    TagFilter:
      name: tag_id
      in: query
      description: Require membership in this exact tag, combined with all other filters using AND.
      schema: { $ref: '#/components/schemas/Uuid' }
    CollectionFilter:
      name: collection_id
      in: query
      description: Require membership in this exact collection, combined with all other filters using AND.
      schema: { $ref: '#/components/schemas/Uuid' }
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string, minLength: 32, maxLength: 256 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: >-
        Retry key scoped to the local user and operation. For anonymous access,
        replay is additionally isolated to the browser session and configured role.
      schema: { type: string, minLength: 16, maxLength: 128 }
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string, minLength: 1, maxLength: 512 }
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
    BoxCode:
      name: box_code
      in: path
      required: true
      schema: { $ref: '#/components/schemas/BoxCode' }
...
  responses:
...
    Unauthorized:
      description: Authentication is required or no longer valid.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Forbidden:
      description: The current user is not authorized.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    NotFound:
      description: The requested resource was not found.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Conflict:
      description: Current lifecycle or domain state conflicts with the request.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
...
    PayloadTooLarge:
      description: Request or decoded image exceeds configured limits.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    UnsupportedMediaType:
      description: Uploaded bytes are not a supported image type.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    ValidationFailed:
      description: One or more request fields failed validation.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
...
    InsufficientStorage:
      description: Local disk reserve prevents this operation.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
  schemas:
...
    DateTime:
      type: string
      format: date-time
...
    SessionView:
      type: object
      additionalProperties: false
      required: [user, anonymous, csrf_token, expires_at, capabilities]
      properties:
        user: { $ref: '#/components/schemas/UserView' }
        anonymous:
          type: boolean
          description: True for the reserved anonymous principal; its role is configured server-side.
        csrf_token: { type: string }
        expires_at: { $ref: '#/components/schemas/DateTime' }
        capabilities:
          type: array
          uniqueItems: true
          items: { type: string }
    UserView:
      type: object
      additionalProperties: false
      required: [id, username, display_name, role, status, version, created_at, updated_at, is_system_admin, local_password]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        username: { type: string }
        is_system_admin: { type: boolean }
        local_password: { type: boolean }
        display_name: { type: string }
        role: { $ref: '#/components/schemas/Role' }
        status: { $ref: '#/components/schemas/UserStatus' }
        version: { type: integer, minimum: 1 }
        created_at: { $ref: '#/components/schemas/DateTime' }
        updated_at: { $ref: '#/components/schemas/DateTime' }
    UserCreate:
      type: object
      additionalProperties: false
      required: [username, display_name, role, password]
      properties:
        username: { type: string, minLength: 3, maxLength: 64 }
        display_name: { type: string, minLength: 1, maxLength: 120 }
        role: { $ref: '#/components/schemas/Role' }
        password: { type: string, minLength: 12, maxLength: 1024 }
...
    BoxCreate:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        description_markdown: { type: string, maxLength: 65536, default: '' }
        tags: { $ref: '#/components/schemas/TagNames' }
        collection_ids: { $ref: '#/components/schemas/CollectionIds' }
...
    BoxDetail:
      allOf:
        - $ref: '#/components/schemas/BoxSummary'
        - type: object
          additionalProperties: true
          required: [description, images, items]
          properties:
            description: { $ref: '#/components/schemas/MarkdownView' }
            images:
              type: array
              items: { $ref: '#/components/schemas/ImageView' }
            items:
              type: array
              items: { $ref: '#/components/schemas/InventoryItemView' }
    BoxPage:
      type: object
      additionalProperties: false
      required: [items, page]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/BoxSummary' }
        page: { $ref: '#/components/schemas/PageInfo' }
...
    InventoryItemCreate:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 160 }
        quantity: { anyOf: [{ $ref: '#/components/schemas/Quantity' }, { type: 'null' }] }
        unit: { type: [string, 'null'], minLength: 1, maxLength: 24 }
        notes_markdown: { type: string, maxLength: 16384, default: '' }
...